backend/pyproject.toml
[project]
name = "seo-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Gaurav Vijay Jadhav",email = "gaurav.vjadhav01@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "fastapi[all] (>=0.116.1,<0.117.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.5,<5.0.0)",
    "lxml (>=6.0.1,<7.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis[hiredis] (>=6.4.0,<7.0.0)",
    "anthropic (>=0.66.0,<0.67.0)",
    "openai (>=1.107.0,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-openai (>=0.3.32,<0.4.0)",
    "langchain-anthropic (>=0.3.19,<0.4.0)",
    "langgraph (>=0.6.7,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
